fixup! JAMES-2266 Allow moving messages when their mailbox no more exist
