--- mailbox/store/src/main/java/org/apache/james/mailbox/store/StoreMessageIdManagerNoChecks.java
@@ -1,61 +1,0 @@
-/****************************************************************
- * Licensed to the Apache Software Foundation (ASF) under one   *
- * or more contributor license agreements.  See the NOTICE file *
- * distributed with this work for additional information        *
- * regarding copyright ownership.  The ASF licenses this file   *
- * to you under the Apache License, Version 2.0 (the            *
- * "License"); you may not use this file except in compliance   *
- * with the License.  You may obtain a copy of the License at   *
- *                                                              *
- *   http://www.apache.org/licenses/LICENSE-2.0                 *
- *                                                              *
- * Unless required by applicable law or agreed to in writing,   *
- * software distributed under the License is distributed on an  *
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY       *
- * KIND, either express or implied.  See the License for the    *
- * specific language governing permissions and limitations      *
- * under the License.                                           *
- ****************************************************************/
-
-package org.apache.james.mailbox.store;
-
-import static org.apache.james.mailbox.store.StoreMessageIdManager.toMailboxIds;
-
-import java.util.List;
-
-import javax.inject.Inject;
-
-import org.apache.james.mailbox.MailboxSession;
-import org.apache.james.mailbox.exception.MailboxException;
-import org.apache.james.mailbox.model.MailboxId;
-import org.apache.james.mailbox.model.MessageId;
-import org.apache.james.mailbox.store.StoreMessageIdManager.MessageMoves;
-import org.apache.james.mailbox.store.mail.MessageIdMapper;
-import org.apache.james.mailbox.store.mail.MessageMapper;
-import org.apache.james.mailbox.store.mail.model.MailboxMessage;
-
-import com.google.common.collect.ImmutableList;
-
-public class StoreMessageIdManagerNoChecks {
-
-    private final MailboxSessionMapperFactory mailboxSessionMapperFactory;
-    private final StoreMessageIdManager messageIdManager;
-
-    @Inject
-    public StoreMessageIdManagerNoChecks(MailboxSessionMapperFactory mailboxSessionMapperFactory, StoreMessageIdManager messageIdManager) {
-        this.mailboxSessionMapperFactory = mailboxSessionMapperFactory;
-        this.messageIdManager = messageIdManager;
-    }
-
-    public void setInMailboxesNoCheck(MessageId messageId, MailboxId targetMailboxId, MailboxSession mailboxSession) throws MailboxException {
-        MessageIdMapper messageIdMapper = mailboxSessionMapperFactory.getMessageIdMapper(mailboxSession);
-        List<MailboxMessage> currentMailboxMessages = messageIdMapper.find(ImmutableList.of(messageId), MessageMapper.FetchType.Full);
-
-        MessageMoves messageMoves = new MessageMoves(toMailboxIds(currentMailboxMessages), ImmutableList.of(targetMailboxId));
-
-        if (messageMoves.isChange()) {
-            messageIdManager.applyMessageMoveNoMailboxChecks(mailboxSession, currentMailboxMessages, messageMoves);
-        }
-    }
-
-}
